+++ mod/twitch/src/main/java/tv/twitch/android/feature/theatre/metadata/MetadataViewEvent.java
@@ -1,0 +1,7 @@
+package tv.twitch.android.feature.theatre.metadata;
+
+public class MetadataViewEvent {
+    public static final class SubscribeButtonClicked extends MetadataViewEvent {
+        public tv.twitch.android.models.channel.ChannelModel channelModel;
+    }
+}